+++ src/Diagnostic/DiagnosticProbes.h
@@ -134,6 +134,9 @@ private:
     
     //! patch size
     std::vector<double> patch_length;
+    
+    //! Datatype for writing to HDF5 file
+    hid_t file_datatype_;
 };